app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
firebase_admin==6.4.0
pandas==2.0.1
Pillow==9.5.0
tensorflow== 2.16.1
joblib==1.2.0
python-multipart==0.0.9
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
Pillow>=9.4.0
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.232
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1
wheel>=0.38.0
pathlib==1.0.1
yolov5==7.0.13